Version 0.71
Amount of Oil currently available is now shown on the shipyard missions map.
The Potential Counters section of the follower class/specialization tooltip now lists possible counter combinations, along with the number of active and inactive followers with those combinations, as well as the total number of recruited followers which could potentially have that combination.

Signed-off-by: Konctantin <gawrilyako@gmail.com>